[ETCM-126] add eth_pendingTransactions and remove qa_getPendingTransactions
