+++ src/main/scala/io/iohk/ethereum/jsonrpc/QAJsonMethodsImplicits.scala
@@ -1,15 +1,11 @@
 package io.iohk.ethereum.jsonrpc
 
-import io.iohk.ethereum.jsonrpc.JsonRpcController.JsonDecoder.NoParamsDecoder
-import io.iohk.ethereum.jsonrpc.JsonRpcController.{Codec, JsonEncoder}
+import io.iohk.ethereum.jsonrpc.JsonRpcController.Codec
 import io.iohk.ethereum.jsonrpc.JsonRpcErrors.InvalidParams
 import io.iohk.ethereum.jsonrpc.QAService.{
-  GetPendingTransactionsRequest,
-  GetPendingTransactionsResponse,
   MineBlocksRequest,
   MineBlocksResponse
 }
-import io.iohk.ethereum.transactions.PendingTransactionsManager.PendingTransaction
 import org.json4s.JsonAST._
 
 object QAJsonMethodsImplicits extends JsonMethodsImplicits {
@@ -35,13 +31,5 @@ object QAJsonMethodsImplicits extends JsonMethodsImplicits {
         "responseType" -> JString(t.responseType.entryName),
         "message" -> t.message.fold[JValue](JNull)(JString)
       )
-    }
-
-  implicit val qa_getPendingTransactions: Codec[GetPendingTransactionsRequest, GetPendingTransactionsResponse] =
-    new NoParamsDecoder(GetPendingTransactionsRequest()) with JsonEncoder[GetPendingTransactionsResponse] {
-      def encodeJson(t: GetPendingTransactionsResponse): JValue =
-        JArray(t.pendingTransactions.toList.map { pendingTx: PendingTransaction =>
-          encodeAsHex(pendingTx.stx.tx.hash)
-        })
     }
 }
