+++ src/main/scala/io/iohk/ethereum/jsonrpc/QAService.scala
@@ -1,32 +1,25 @@
 package io.iohk.ethereum.jsonrpc
 
-import akka.actor.ActorRef
+import akka.util.ByteString
 import akka.util.ByteString
 import cats.implicits._
 import enumeratum._
 import io.iohk.ethereum.consensus._
 import io.iohk.ethereum.consensus.ethash.MinerResponses._
 import io.iohk.ethereum.consensus.ethash.MockedMinerProtocol.MineBlocks
-import io.iohk.ethereum.consensus.ethash.{MinerResponse, MinerResponses, TransactionPicker}
+import io.iohk.ethereum.consensus.ethash.{MinerResponse, MinerResponses}
 import io.iohk.ethereum.jsonrpc.QAService.MineBlocksResponse.MinerResponseType
 import io.iohk.ethereum.jsonrpc.QAService.{
-  GetPendingTransactionsRequest,
-  GetPendingTransactionsResponse,
   MineBlocksRequest,
   MineBlocksResponse
 }
-import io.iohk.ethereum.transactions.PendingTransactionsManager.PendingTransaction
 import io.iohk.ethereum.utils.Logger
 import monix.execution.Scheduler.Implicits.global
 import mouse.all._
-import scala.concurrent.duration.FiniteDuration
 
 class QAService(
-    consensus: Consensus,
-    val pendingTransactionsManager: ActorRef,
-    val getTransactionFromPoolTimeout: FiniteDuration
-) extends Logger
-    with TransactionPicker {
+    consensus: Consensus
+) extends Logger {
 
   /**
     * qa_mineBlocks that instructs mocked miner to mine given number of blocks
@@ -44,16 +37,6 @@ class QAService(
           Left(JsonRpcErrors.InternalError)
       }
   }
-
-  /**
-    * qa_getPendingTransactions that returns all pending transactions from the mempool
-    *
-    * @return all pending transactions from the mempool
-    */
-  def getPendingTransactions(req: GetPendingTransactionsRequest): ServiceResponse[GetPendingTransactionsResponse] =
-    getTransactionsFromPool.map { resp =>
-      Right(GetPendingTransactionsResponse(resp.pendingTransactions))
-    }
 }
 
 object QAService {
@@ -88,7 +71,4 @@ object QAService {
       }
     }
   }
-
-  case class GetPendingTransactionsRequest()
-  case class GetPendingTransactionsResponse(pendingTransactions: Seq[PendingTransaction])
 }
