+++ src/test/scala/io/iohk/ethereum/jsonrpc/QaJRCSpec.scala
@@ -1,21 +1,15 @@
 package io.iohk.ethereum.jsonrpc
 
-import akka.util.ByteString
 import io.iohk.ethereum.NormalPatience
 import io.iohk.ethereum.consensus.ethash.MockedMinerProtocol.MineBlocks
 import io.iohk.ethereum.consensus.ethash.{MinerResponse, MinerResponses}
-import io.iohk.ethereum.crypto.ECDSASignature
 import io.iohk.ethereum.db.storage.AppStateStorage
-import io.iohk.ethereum.domain.{Address, SignedTransactionWithSender, Transaction}
 import io.iohk.ethereum.jsonrpc.JsonRpcController.JsonRpcConfig
 import io.iohk.ethereum.jsonrpc.QAService.MineBlocksResponse.MinerResponseType._
 import io.iohk.ethereum.jsonrpc.QAService.{
-  GetPendingTransactionsRequest,
-  GetPendingTransactionsResponse,
   MineBlocksRequest,
   MineBlocksResponse
 }
-import io.iohk.ethereum.transactions.PendingTransactionsManager.PendingTransaction
 import io.iohk.ethereum.utils.Config
 import org.json4s.JsonAST._
 import org.scalamock.scalatest.MockFactory
@@ -80,51 +74,6 @@ class QaJRCSpec extends AnyWordSpec with Matchers with ScalaFutures with NormalP
         response should haveError(JsonRpcErrors.InternalError)
       }
     }
-
-    "request pending transactions and return valid response" when {
-      "mempool is empty" in new TestSetup {
-        (qaService.getPendingTransactions _)
-          .expects(getPendingTransactionReq)
-          .returning(Future.successful(Right(GetPendingTransactionsResponse(List()))))
-
-        val response: JsonRpcResponse = jsonRpcController.handleRequest(getPendingTransactionsRpcRequest).futureValue
-
-        response should haveResult(JArray(List()))
-      }
-
-      "mempool has transactions" in new TestSetup {
-        val transactions = (0 to 1).map(_ => {
-          val fakeTransaction = SignedTransactionWithSender(
-            Transaction(
-              nonce = 0,
-              gasPrice = 123,
-              gasLimit = 123,
-              receivingAddress = Address("0x1234"),
-              value = 0,
-              payload = ByteString()
-            ),
-            signature = ECDSASignature(0, 0, 0.toByte),
-            sender = Address("0x1234")
-          )
-          PendingTransaction(fakeTransaction, System.currentTimeMillis)
-        })
-        (qaService.getPendingTransactions _)
-          .expects(getPendingTransactionReq)
-          .returning(Future.successful(Right(GetPendingTransactionsResponse(transactions))))
-
-        val response: JsonRpcResponse = jsonRpcController.handleRequest(getPendingTransactionsRpcRequest).futureValue
-
-        val result = JArray(
-          transactions
-            .map(tx => {
-              encodeAsHex(tx.stx.tx.hash)
-            })
-            .toList
-        )
-
-        response should haveResult(result)
-      }
-    }
   }
 
   trait TestSetup extends MockFactory with JRCMatchers {
@@ -142,7 +91,6 @@ class QaJRCSpec extends AnyWordSpec with Matchers with ScalaFutures with NormalP
       new JsonRpcController(web3Service, netService, ethService, personalService, None, debugService, qaService, config)
 
     val mineBlocksReq = MineBlocksRequest(1, true, None)
-    val getPendingTransactionReq = GetPendingTransactionsRequest()
 
     val mineBlocksRpcRequest = JsonRpcRequest(
       "2.0",
@@ -153,17 +101,6 @@ class QaJRCSpec extends AnyWordSpec with Matchers with ScalaFutures with NormalP
             JInt(1),
             JBool(true)
           )
-        )
-      ),
-      Some(JInt(1))
-    )
-
-    val getPendingTransactionsRpcRequest = JsonRpcRequest(
-      "2.0",
-      "qa_getPendingTransactions",
-      Some(
-        JArray(
-          List()
         )
       ),
       Some(JInt(1))
